fix: move auth.ts file to app/actions and use fetchServer for requests
